+++ doc/release_notes.rst
@@ -1,6 +1,12 @@
 Release Notes
 =============
 
+Upcoming Version
+----------------
+
+* The return type of ``coord_dims`` for expressions and constraints was changed from set to tuple to align with the xarray convention.
+* The printout of transposed expressions and constraints was fixed.
+
 
 Version 0.3.4
 -------------